requirements.txt
streamlit
pandas
pymongo
langchain
sentence-transformers
google-generativeai
python-dotenv
certifi
openai